server resources packaging
