+++ Robust.Packaging/RobustClientPackaging.cs
@@ -35,29 +35,6 @@ public sealed class RobustClientPackaging
             .ToHashSet();
 
         await RobustSharedPackaging.DoResourceCopy(Path.Combine(contentDir, "Resources"), pass, ignoreSet, cancel: cancel);
-
-        var manifestPath = Path.Combine(contentDir, "Resources", "manifest.yml");
-        var manifest = ResourceManifestData.LoadFromFile(manifestPath); // load from disk no VFS
-
-        if (manifest.ModularResources != null)
-        {
-            foreach (var (vfsPath,diskName) in manifest.ModularResources)
-            {
-                var modPath = Path.Combine(contentDir, diskName);
-                if (!Directory.Exists(modPath))
-                    continue;
-                var files = Directory.EnumerateFiles(modPath, "*", SearchOption.AllDirectories);
-                foreach (var file in files)
-                {
-                    var filename = Path.GetFileName(file);
-                    if (ignoreSet.Contains(filename)) continue;
-                    var relative = Path.GetRelativePath(modPath, file);
-                    var zipRoot = vfsPath.TrimStart('/');
-                    var targetPath = Path.Combine(zipRoot, relative).Replace('\\', '/');
-                    await using var stream = File.OpenRead(file);
-                    pass.InjectFileFromDisk(targetPath, file);
-                }
-            }
-        }
+        await RobustSharedPackaging.DoModularResourceCopy(contentDir, pass, ignoreSet);
     }
 }
